feat: 컬럼 추가로 인한 Manage 모듈 수정

+++ core/database/src/main/java/com/project/database/model/HoldIngredientEntity.kt
@@ -20,7 +20,7 @@ import java.time.LocalDate
 data class HoldIngredientEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     @ColumnInfo(name = "ingredient_id") val ingredientId: Int,
-    val count: Int,
+    val count: Double,
     val enterDate: LocalDate,
     val expirationDate: LocalDate,
 )
