+++ domain/ingredient/src/main/java/com/project/ingredient/repository/HoldIngredientRepository.kt
@@ -7,5 +7,5 @@ interface HoldIngredientRepository {
     fun getAllHoldIngredient(): Flow<List<Ingredient>>
     suspend fun deleteHoldIngredientByIds(ids: List<Int>)
     suspend fun getHoldIngredientById(id: Int): Ingredient
-    suspend fun updateHoldIngredientCount(id: Int, count: Int)
+    suspend fun updateHoldIngredientCount(id: Int, count: Double)
 }
